Allow machine and contexts to have different type parameters
The code generation presently assumes that the state machine and its
context, if present, will share the same type parameters. This works
fine in the common case where neither the machine nor the context have
any type parameters, but breaks down when, say, the state machine has
type parameters but the context does not.

Require that the #[state_machine_future(context = ...)] annotation
specify explicitly what type parameters exist on the context, and with
what bounds. This allows code generation to propagate the type
parameters correctly. For example, if only the context has type
parameter T, the generated future must be generic over T, even though
the state machine enum will not be.

Note that this is a breaking change for state machines where the machine
and the context have exactly the same type parameters--for an example of
this, see the call_to_context test--as the context annotation will need
to be updated to specify the type parameter. I don't see a solution that
avoids this, unfortunately, but on the bright side the common case where
no type parameters are used is unaffected.